Add toolbar shortcuts and icons
