--- Analyze/Microscope Measurement Tools/MScopeCals - custom function example/JEOL SEM - etched sidewall 01.txt
@@ -1,83 +1,0 @@
-$SEM_DATA_VERSION 1
-$CM_SIGNAL_NAME SEI
-$$SM_ADD_IMAGE 
-$$SM_ADD_CHECKED_CH 1
-$$ADD_CH1_SIGNAL SEI
-$$ADD_CH2_SIGNAL SEI
-$$ADD_CH3_SIGNAL SEI
-$$ADD_CH4_SIGNAL SEI
-$CM_CONTRAST 2390
-$CM_BRIGHTNESS 544
-$$SM_MIX_IMAGE 
-$CM_ACCEL_VOLT 1.00
-$SM_ARRIVAL_EMI 0.00
-$SM_GB_GUN_VOLT 3.00
-$SM_GB_BIAS_VOLT 2.00
-$$SM_WD 2.77
-$CM_MAG 50000
-$$SM_COLUMN_MODE GB_HIGH
-$$SM_OL_APERTURE 9
-$$SM_CL_COARSE 14
-$$SM_CL_FINE 25
-$$SM_GUN_ALI_TILT 0 0
-$$SM_CL_STIG 0 0
-$$SM_OL_STIG 68 162
-$$SM_OL_STIG_CENTER 0 0 0 0
-$$SM_OLAP_ALI 0 0
-$$SM_PROBE_NO 6
-$$SM_PROBE_FINE 25
-$$SM_ACL_ALI -2 -1
-$$SM_COLUMN_CENTER 0 0
-$$SM_EF_ALI 0 0
-$$SM_SCAN_ROTATION 298.80
-$$SM_WD_CORRECT 0
-$$SM_INTEG_NO 1
-$$CM_SCAN_MODE 3
-$$CM_SCAN_SPEED 9
-$$CM_INSTRUMENT_SCAN_SPEED 9
-$$SM_R_FILTER_MODE 0
-$$SM_SEI_COLLECTOR 0
-$$SM_R_FILTER_INDEX 101
-$$SM_DET_SE1_MODE_NUMBER 0
-$$SM_DET_TTL_MODE_NUMBER 0
-$$SM_DET_SE2_MODE_NUMBER 0
-$$SM_TTL_COLLECTOR_VOLTAGE 0.00
-$$SM_SED_FILTER_MODE 0
-$$SM_UED_FILTER_MODE 0
-$$SM_UED_FILTER_VALUE 0.00
-$$SM_ACB_MODE 1
-$$SM_ACB_CONTRAST 3.36
-$$SM_ACB_BRIGHTNESS -5.54
-$CM_SIGNAL SEI
-$CM_STAGE_POS -0.169 1.289 7.4 360.0 0.0 0
-$SM_EMI_CURRENT 46.20
-$$SM_PROBE_CURRENT 0.00
-$SM_PENNING_VAC 9.634e-005
-$SM_PRESSURE_AT_LOW_VACUUM 0.000
-$SM_VACUUM_MODE_FOR_SPECIMEN_CHAMBER 0
-$CM_FORMAT JEOL-SEM
-$CM_VERSION 1.0
-$CM_COMMENT 
-$CM_DATE 2010/10/01
-$CM_TIME 08:44:19
-$CM_OPERATOR Blumenthal
-$CM_INSTRUMENT 7600F
-$$SM_MICRON_BAR 53
-$$SM_MICRON_MARKER 100nm
-$CM_FRAME_SIZE 0 0
-$CM_TITLE JEOL SEM - etched sidewall 01
-$CM_LABEL OCPN
-$CM_OFFSET 0 0
-$CM_FULL_SIZE 1280 960
-$AN_CURSOR_MEAS_TYPE 0
-$AN_EXPORT_MODE 1
-$CM_COLOR_MODE 0
-$AN_ANNOTATION_MODE 1
-$AN_CURSOR_MEAS $AN_LINE%0 $AN_LINE%1 $AN_LINE%2 $AN_LINE%3 $AN_TEXT%0 $AN_TEXT%1 $AN_TEXT%2
-$AN_LINE%0 440.00000000 0.00000000 440.00000000 959.00000000 65280 1.00000000 0 0
-$AN_LINE%1 840.00000000 0.00000000 840.00000000 959.00000000 65280 1.00000000 0 0
-$AN_LINE%2 0.00000000 280.00000000 1279.00000000 280.00000000 65280 1.00000000 0 0
-$AN_LINE%3 0.00000000 680.00000000 1279.00000000 680.00000000 65280 1.00000000 0 0
-$AN_TEXT%0 5.00000000 874.00000000 215.00000000 910.00000000 "X:" "Courier New" 0 32.00000000 16777215 65793 1 0
-$AN_TEXT%1 220.00000000 874.00000000 430.00000000 910.00000000 "Y:" "Courier New" 0 32.00000000 16777215 65793 1 0
-$AN_TEXT%2 435.00000000 874.00000000 645.00000000 910.00000000 "D:" "Courier New" 0 32.00000000 16777215 65793 1 0
